HBASE-21322 Add a scheduleServerCrashProcedure() API to HbckService
Adds server version check so we fail fast and complain if remote
server does not support feature.

Refactor so we test server version before we use feature/command.

Renamed serverCrashProcedures as scheduleRecoveries

Add a new Version class that has Version math and check in it.

Add tests and refactor of tests because HBCK2 format changed.

Signed-off-by: Wellington Chevreuil <wchevreuil@apache.org>