HBASE-25297 [HBCK2] Regenerate missing table descriptors by hbck2
We just have been in a situation, when we saw many ServerCrashProcedure to
be blocked, because of the AssignProcedure is blocked because of the missing
.tabledesc files.

java.io.IOException: Missing table descriptor for dd0676f57bdbff5d04ab735b7daf5e9b

In our case it was OK for us to get rid of these tables and we were able to
use setRegionState to move all these regions to FAILED_OPEN state, then disable
and drop the tables. But this took a lot of time and we might not always have
the option to drop these tables.

HBCK 1 had a functionality (fixTableOrphans) to regenerate the table descriptors
according to the memory cache or hdfs table directory structure. In this patch
I implemented the same logic for HBCK 2. I created tests for the new feature and
also tested it with a real HBase 2.6.1 cluster.